finished make attendance

+++ src/main/java/com/itcube/journal/repos/CourseRepo.java
@@ -3,5 +3,5 @@ package com.itcube.journal.repos;
 import com.itcube.journal.model.Course;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CourseRepo extends CrudRepository<Course, Integer> {
+public interface CourseRepo extends CrudRepository<Course, Long> {
 }
